Move definition code into __init__.py
